% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_optimal.r
\name{CV_optimal}
\alias{CV_optimal}
\title{Cross Validation optimal revised}
\usage{
CV_optimal(V, h, k)
}
\arguments{
\item{V}{A value to return}

\item{h}{the list of bandwidth parameters to test in cross validation}

\item{k}{kernel function}
}
\value{
Returns the value of
}
\description{
\code{CV_optimal_revised} does something.
}
\examples{

#myfunction(1) # returns 1

}
