% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_binomial.R
\name{discrete_binomial}
\alias{discrete_binomial}
\title{Discrete binomial kernel}
\usage{
discrete_binomial(x, z, h)
}
\arguments{
\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter) which should match the
condition 0<= h < 1}
}
\value{
Returns the value of the associated kernel function according to the
target x and the bandwidth h.
}
\description{
Discrete binomial kernel
}
\examples{
# Basic usage of discrete_binomial() to compute a Discrete Binomial Kernel
discrete_binomial(x = 25, z = 1:50, h = 0.5)
}
