% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_epanech.r
\name{discrete_epanech}
\alias{discrete_epanech}
\title{Discrete Epanechnikov kernel}
\usage{
discrete_epanech(x, z, h)
}
\arguments{
\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter)}
}
\value{
Returns the value of the associated kernel function according to the
target x and the bandwidth h.
}
\description{
Discrete Epanechnikov kernel
}
\examples{
# Basic usage of discrete_epanech() to compute a discrete Epanechnikov kernel
discrete_epanech(x = 25, z = 1:50, h = 20)

}
