% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_kernel.r
\name{discrete_kernel}
\alias{discrete_kernel}
\title{Discrete kernel function}
\usage{
discrete_kernel(kernel = c("optimal", "triang", "epanech"), x, z, h, k = NULL)
}
\arguments{
\item{kernel}{the type of kernel. Currently supported kernels are limited to:
"optimal", "triang" and "epanech"}

\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter)}

\item{k}{Optional: the integer (positive) parameter that defined the support
of the kernel function (corresponds to parameter 'a' for triangular
kernel). It is only used for optimal and triangular kernel}
}
\value{
Returns the value of the associated kernel function
}
\description{
Discrete kernel function
}
\examples{
discrete_kernel(kernel="triang", x = 50, z = 1:100, h = 10, a = 5)
}
\seealso{
\code{\link[=discrete_optimal(), discrete_triang(), discrete_epanech]{discrete_optimal(), discrete_triang(), discrete_epanech()}}  which
this function wraps.
}
