% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteEpanech.r
\name{DiscreteEpanech}
\alias{DiscreteEpanech}
\title{Title}
\usage{
DiscreteEpanech(x, z, h)
}
\arguments{
\item{z}{vector of observations}

\item{h}{kernel bandwidth (or smoothing parameter) is defined as a positive integer}
}
\value{
Returns the value of the associated kernel function at t according to the target and the bandwidth.
}
\description{
Title
}
