% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteOptimal.r
\name{DiscreteOptimal}
\alias{DiscreteOptimal}
\title{Discrete optimal kernel}
\usage{
DiscreteOptimal(x, z, h, k)
}
\arguments{
\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter), which should match the condition (3 / 5) * (1 - 1 / k)) < h < 1}

\item{k}{the integer (positive) parameter that defined the support of the kernel function}
}
\value{
Returns the value of the associated kernel function according to the target x and the bandwidth h.
}
\description{
Discrete optimal kernel
}
\examples{
DiscreteOptimal(x = 50, z = 1:100, h = 0.6, k = 10)

}
