% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteTriang.r
\name{DiscreteTriang}
\alias{DiscreteTriang}
\title{Discrete triangular kernel}
\usage{
DiscreteTriang(x, z, h, a)
}
\arguments{
\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter)}

\item{a}{the integer (positive) parameter that defined the support of the kernel function}
}
\value{
Returns the value of the associated kernel function according to the target x and the bandwidth h.
}
\description{
Discrete triangular kernel
}
\examples{
# basic usage of DiscreteTriang to compute a Discrete triangular kernel
DiscreteTriang(x = 50, z = 1:100, h = 10, a = 5)

}
