% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_kernel.r
\name{estim_kernel}
\alias{estim_kernel}
\title{Estimator for discrete optimal kernel}
\usage{
estim_kernel(kernel = c("optimal", "triang", "epanech"), x, h, v, k = NULL)
}
\arguments{
\item{x}{the target point at which the density is calculated.}

\item{h}{the bandwidth (or smoothing parameter), which should match the
condition (3 / 5) * (1 - 1 / k)) < h < 1}

\item{v}{???}

\item{k}{the integer (positive) parameter that defined the support of
the kernel function}
}
\value{
The estimated optimal kernel
}
\description{
Estimator for discrete optimal kernel
}
