% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Fish dataset from the SIMTAP project}
\format{
\subsection{\code{fish_data}}{

A data frame with 200 rows and 2 columns:
\describe{
\item{length}{Length in mm}
\item{weight}{Weight in dg}
...
}
}
}
\source{
\url{https://www.simtap.eu}
}
\usage{
fish_data
}
\description{
The SIMTAP project (Self-sufficient Integrated Multi-Trophic AquaPonic
systems)  aimed to develop sustainable aquaculture production system
that, in particular, contributes to reduce fish feed inputs and
resources consumption. We used data from an experiment in which gilthead
seabream (sparus aurata) were stocked in 1.6 m3 tanks at a density
of 1.5 kg·m-3 .
Fish were reared for 46 days in a single recirculating aquaculture system
composed of three rearing tanks. At the beginning of the experiment, a number
n =200 of the fish were individually weighed (dg), and their length
at the caudal fork (mm) was measured.
}
\keyword{datasets}
