% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_optimal.r
\name{cv_optimal}
\alias{cv_optimal}
\title{Cross Validation optimal}
\usage{
cv_optimal(v, h, k)
}
\arguments{
\item{v}{A value to return}

\item{h}{the list of bandwidth parameters to test in cross validation}

\item{k}{kernel function}
}
\value{
Returns the optimal bandwith value
#'
}
\description{
\code{cv_optimal} does something.
}
\examples{

# myfunction(1) # returns 1

}
