% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteEpanech.r
\name{DiscreteEpanech}
\alias{DiscreteEpanech}
\title{Title}
\usage{
DiscreteEpanech(x, z, h)
}
\arguments{
\item{x}{the target point at which the density is calculated}

\item{z}{the vector of observations}

\item{h}{the bandwidth (or smoothing parameter)}
}
\value{
Returns the value of the associated kernel function according to the target x and the bandwidth h.
}
\description{
Title
}
\examples{
# basic usage of DiscreteEpanech to compute a discrete Epanechnikov kernel
DiscreteEpanech(x = 50, z = 1:100, h = 10)

}
