% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_kernel.r
\name{estim_kernel}
\alias{estim_kernel}
\title{Estimator for discrete optimal kernel}
\usage{
estim_kernel(kernel = c("optimal", "triang", "epanech"), x, h, v, k = NULL)
}
\arguments{
\item{kernel}{the type of kernel. Currently supported kernels are limited to:
"optimal", "triang" and "epanech"}

\item{x}{the list of target points at which the density is calculated  TODO : à valider}

\item{h}{the bandwidth (or smoothing parameter)}

\item{v}{the vector of observations TODO : Renommer "z" ???}

\item{k}{Optional: the integer (positive) parameter that defined the support
of the kernel function (corresponds to parameter 'a' for triangular
kernel). It is only used for optimal and triangular kernel}
}
\value{
The estimated kernel
}
\description{
Estimator for discrete optimal kernel
}
\examples{
fn_opt_k <- estim_kernel(kernel=kernel, x=count_values, h=hcv, v=fish_weights, k=k)
}
